Introduced auth.
Works OK-isch but code is not fool proof and not all errors are caught (yet)
This is alpha version
Search for xxx in ocde and fill in your call back domain, client_id and other secrets.

In essence header is used and instead of bearer_id, the token is used. This works.